# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tourist Route Management API",
        "description": "API documentation for the Tourist Route project",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/itineraries": {
            "get": {
                "tags": [
                    "Itinerarie"
                ],
                "summary": "List itineraries",
                "operationId": "5450dc8a5c7fc4f8fd5725fb376b9d06",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Itinerarie"
                ],
                "summary": "Create itinerarie",
                "operationId": "08d26fd9afc5118415c3e8091f549204",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "duration_from",
                                    "duration_to",
                                    "destinations"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "duration_from": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "duration_to": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "image": {
                                        "type": "string",
                                        "nullable": true,
                                        "maxLength": 255
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "destinations": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/itineraries/filter": {
            "get": {
                "tags": [
                    "Itinerarie"
                ],
                "summary": "Filter itineraries",
                "operationId": "7ab9d95b35f5f83d2522d496ef50b30f",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "duration_from": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "duration_to": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/itineraries/{itinerarie}": {
            "get": {
                "tags": [
                    "Itinerarie"
                ],
                "summary": "Show itinerarie",
                "operationId": "ca13e12b6a34569090034933c6a39a7e",
                "parameters": [
                    {
                        "name": "itinerarie",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Itinerarie"
                ],
                "summary": "Update itinerarie",
                "operationId": "0de281e04c0a0741655aa5a1e6d207ed",
                "parameters": [
                    {
                        "name": "itinerarie",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "duration_from": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "duration_to": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "image": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Itinerarie"
                ],
                "summary": "Delete itinerarie",
                "operationId": "6969091c3d52a4dae4eb20138503020c",
                "parameters": [
                    {
                        "name": "itinerarie",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}